// apps/admin-web/src/.generated/api/model/webhookConfigDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */
import type { WebhookConfigDtoType } from './webhookConfigDtoType';

export interface WebhookConfigDto {
  /** 동일 문의 미응답 알림 재발송 간격 (분) */
  cooldownMinutes: number;
  /** 알림 연동 활성화 여부 */
  enabled: boolean;
  /** 웹훅 전송 채널 종류 */
  type: WebhookConfigDtoType;
  /** 알림 수신 웹훅 URL */
  webhookUrl: string;
}