// apps/admin-web/src/.generated/api/model/webhookConfigDtoType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */

/**
 * 웹훅 전송 채널 종류
 */
export type WebhookConfigDtoType = typeof WebhookConfigDtoType[keyof typeof WebhookConfigDtoType];


export const WebhookConfigDtoType = {
  SLACK: 'SLACK',
  DISCORD: 'DISCORD',
  CHANNEL_TALK: 'CHANNEL_TALK',
  TEAMS: 'TEAMS',
} as const;